[#126] Coordinator 변경
