+++ Wasap/Wasap/Features/WifiAutoConnect/ViewModel/ConnectingViewModel.swift
@@ -55,11 +55,9 @@ public class ConnectingViewModel: BaseViewModel {
                     default:
                         self.coordinatorController?.performFinish(to: .finishWithError)
                     }
-
                 } else {
                     self.coordinatorController?.performFinish(to: .finishWithError)
                 }
-
                 Log.error("Wi-Fi 연결 중 에러 발생: \(error.localizedDescription)")
             }
             .disposed(by: disposeBag)
