+++ Wasap/Wasap/Global/Base/BaseViewController.swift
@@ -28,7 +28,6 @@ public class RxBaseViewController<VM: BaseViewModelProtocol>: UIViewController {
 
     public override func viewDidLoad() {
         super.viewDidLoad()
-        
         view.backgroundColor = .systemBackground
     }
